requirements.txt
torch==2.1.0
torchvision==0.16.0
fastapi==0.104.1
uvicorn==0.24.0
mlflow==2.9.2
prometheus-client==0.19.0
Pillow==10.1.0
scikit-learn==1.3.2
numpy==1.26.2
python-multipart==0.0.6
matplotlib==3.8.2
dvc==3.30.1
pytest==7.4.3
httpx==0.25.2
